pyproject.toml
[tool.poetry]
name = "iydon"
version = "2023.4"  # year.iota
description = "Iydon's common tools"
authors = ["Iydon Liang <liangiydon@gmail.com>"]
keywords = ["utilities"]
readme = "README.md"
license = "GPL-3.0-only"
homepage = "https://github.com/iydon/iydon"
repository = "https://github.com/iydon/iydon"
include = []
classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: GNU General Public License v3 (GPLv3)",
    "Natural Language :: Chinese (Simplified)",
    "Natural Language :: English",
    # TODO: https://stackoverflow.com/questions/64138572/pyenv-global-interpreter-not-working-on-windows10
    # "Operating System :: Microsoft :: Windows",
    "Operating System :: POSIX :: Linux",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    # "Programming Language :: Python :: 3.11",
    # "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: User Interfaces",
    "Topic :: Utilities",
    "Typing :: Typed",
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/iydon/iydon/issues"

[tool.poetry.dependencies]
python = ">=3.7,<4.0"
typing-extensions = {version = "^4.5.0", optional = true}

[tool.poetry.extras]
type = ["typing-extensions"]
full = ["typing-extensions"]

[tool.poetry.dev-dependencies]
click = "^8.1.3"
mkdocs = "^1.4.2"
mkdocs-material = "^9.1.4"
mkdocstrings = {extras = ["python"], version = "^0.20.0"}
mypy = "^1.1.1"
nox-poetry = "^1.0.2"
PyGithub = "^1.58.1"
pytest = "^7.2.2"
tqdm = "^4.65.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.script.config]
upgrade = [
    'pyproject.toml',
    'iydon/base/constant.py',
]
